refactor: merge temp-controller to user-controller
