+++ controllers/user-controller.js
@@ -4,7 +4,7 @@ const Op = require('../models').Sequelize.Op
 const { imgurFileHandler } = require('../helpers/file-helper')
 const { User, Followship, Tweet, Like, Reply, sequelize } = require('../models')
 const jwtHelpers = require('../helpers/bearer-token-helper')
-const helper = require('../_helpers')
+const authHelpers = require('../_helpers')
 const jwt = require('jsonwebtoken')
 const {
   putUserCheck,
@@ -118,7 +118,7 @@ const userController = {
         ],
         attributes: { exclude: ['password'] }
       })
-      const followedUsers = helper.getUser(req).Followings
+      const followedUsers = authHelpers.getUser(req).Followings
       user.dataValues.isFollowed = followedUsers.some(fu => fu.id === user.id)
 
       return res.json(user)
@@ -128,7 +128,7 @@ const userController = {
   },
   getTopUsers: async (req, res, next) => {
     try {
-      const user = helper.getUser(req)
+      const user = authHelpers.getUser(req)
       const users = await User.findAll({
         where: {
           id: {
@@ -148,7 +148,7 @@ const userController = {
         limit: 10
       })
 
-      const followedUsers = helper.getUser(req).Followings
+      const followedUsers = authHelpers.getUser(req).Followings
 
       const results = users.map(u => ({
         ...u.toJSON(),
@@ -162,7 +162,7 @@ const userController = {
   },
   getCurrentUser: async (req, res, next) => {
     try {
-      const userId = helper.getUser(req).id
+      const userId = authHelpers.getUser(req).id
 
       const currentUser = await User.findByPk(userId, {
         attributes: [
@@ -189,7 +189,7 @@ const userController = {
 
       const error = new Error()
       const targetUserId = req.params.id
-      const loginUserId = helper.getUser(req).id
+      const loginUserId = authHelpers.getUser(req).id
 
       if (isNaN(targetUserId) || !(await User.findByPk(targetUserId))) {
         error.code = 404
@@ -243,7 +243,7 @@ const userController = {
   putUserSetting: async (req, res, next) => {
     try {
       const error = new Error()
-      const loginUserId = helper.getUser(req).id
+      const loginUserId = authHelpers.getUser(req).id
       let targetUserId = req.params.id
 
 
@@ -294,7 +294,7 @@ const userController = {
     try {
       const error = new Error()
       let targetUserId = req.params.id
-      const loginUserId = helper.getUser(req).id
+      const loginUserId = authHelpers.getUser(req).id
       const DEL_OPERATION_CODE = '-1'
 
       // 測試檔要求
@@ -367,8 +367,296 @@ const userController = {
       error.code = 500
       return next(error)
     }
-  }
+  },
+  // 獲取指定使用者所追隨的使用者清單
+  getFollowings: async (req, res, next) => {
+    try {
+      const error = new Error()
+      const loginUserId = authHelpers.getUser(req).id
+      const targetUserId = req.params.id
 
+      // 找不到使用者可以調閱他/她追隨的使用者清單
+      if (isNaN(targetUserId) || !(await User.findByPk(targetUserId))) {
+        error.code = 404
+        error.message = '對應使用者不存在'
+        return next(error)
+      }
+
+      // 目前使用者調閱他人所追隨的使用者清單
+      // (含暱稱/名稱、帳號、頭像、封面、自我介紹、它所追蹤的人是否也被目前使用者追蹤)
+      // 按追隨紀錄排序，由新至舊
+      const findOption = {
+        include: [
+          {
+            model: User,
+            as: 'Followings',
+            attributes: [
+              ['id', 'followingId'],
+              'name', 'account', 'introduction',
+              'cover', 'avatar',
+              [
+                sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Followships
+                      WHERE followerId = ${loginUserId} 
+                      AND followingId = Followings.id
+                    )
+                `),
+                'isFollowed'
+              ]
+            ]
+          }
+        ],
+        attributes: [],
+        order: [
+          [sequelize.literal('`Followings->Followship`.`createdAt`'), 'DESC']
+        ]
+      }
+      const followingUsers = await User.findByPk(targetUserId, findOption)
+
+      const result = followingUsers.toJSON().Followings
+
+      result.forEach(fu => {
+        fu.isFollowed = Boolean(fu.isFollowed)
+      })
+
+      return res
+        .status(200)
+        .json(result)
+    } catch (error) {
+      // 系統出錯
+      error.code = 500
+      return next(error)
+    }
+  },
+  // 獲取指定使用者的被跟隨之使用者清單
+  getFollowers: async (req, res, next) => {
+    try {
+      const error = new Error()
+      const loginUserId = authHelpers.getUser(req).id
+      const targetUserId = req.params.id
+
+      // 找不到使用者可以調閱他/她的被跟隨之使用者清單
+      if (isNaN(targetUserId) || !(await User.findByPk(targetUserId))) {
+        error.code = 404
+        error.message = '對應使用者不存在'
+        return next(error)
+      }
+
+      // 目前使用者調閱另一個使用者X的跟隨者清單
+      // (含暱稱/名稱、帳號、頭像、封面、自我介紹、它所追蹤的人是否也被目前使用者追蹤)
+      // 按追隨紀錄排序，由新至舊
+      const findOption = {
+        include: [
+          {
+            model: User,
+            as: 'Followers',
+            attributes: [
+              ['id', 'followerId'],
+              'name', 'account', 'introduction',
+              'cover', 'avatar',
+              [
+                sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Followships
+                      WHERE followerId = ${loginUserId} 
+                      AND followingId = Followers.id
+                    )
+                `),
+                'isFollowed'
+              ]
+            ]
+          }
+        ],
+        attributes: [],
+        order: [
+          [sequelize.literal('`Followers->Followship`.`createdAt`'), 'DESC']
+        ]
+      }
+      const followerUsers = await User.findByPk(targetUserId, findOption)
+
+      const result = followerUsers.toJSON().Followers
+
+      result.forEach(fu => {
+        fu.isFollowed = Boolean(fu.isFollowed)
+      })
+
+      return res
+        .status(200)
+        .json(result)
+
+
+    } catch (error) {
+      // 系統出錯
+      error.code = 500
+      return next(error)
+    }
+  },
+  // 獲取對應使用者所喜歡的推文之清單
+  getLikes: async (req, res, next) => {
+    try {
+      const error = new Error()
+      const loginUserId = authHelpers.getUser(req).id
+      const targetUserId = req.params.id
+
+      // 找不到使用者可以調閱他/她的喜歡推文清單
+      if (isNaN(targetUserId) || !(await User.findByPk(targetUserId))) {
+        error.code = 404
+        error.message = '對應使用者不存在'
+        return next(error)
+      }
+
+      // 目前使用者調閱另一個使用者X的喜歡推文清單
+      // (含暱稱/名稱、帳號、頭像、封面、它所喜歡的推文是否也被目前使用者喜歡和回覆)
+      // (時間、推文內容等)
+      // 按喜歡紀錄排序，由新至舊
+      const findOption = {
+        where: {
+          UserId: targetUserId
+        },
+        include: [
+          {
+            model: Tweet,
+            as: 'LikedTweet',
+            attributes: [
+              'id', 'UserId', 'description', 'createdAt',
+              'updatedAt', 'replyCount', 'likeCount',
+              [
+                sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Likes
+                      WHERE UserId = ${loginUserId} 
+                      AND TweetId = LikedTweet.id
+                    )
+                `),
+                'isLiked'
+              ],
+              [
+                sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Replies
+                      WHERE UserId = ${loginUserId} 
+                      AND TweetId = LikedTweet.id
+                    )
+                `),
+                'isReplied'
+              ]
+            ],
+            include: [
+              {
+                model: User,
+                as: 'TweetAuthor',
+                attributes: [
+                  'id', 'account', 'name', 'email',
+                  'avatar', 'cover'
+                ]
+              }
+            ]
+          }
+        ],
+        order: [['createdAt', 'DESC']]
+      }
+      const LikedTweets = await Like.findAll(findOption)
+      const results = LikedTweets.map(lt => {
+        lt = lt.toJSON()
+        lt.LikedTweet.isLiked = Boolean(lt.LikedTweet.isLiked)
+        lt.LikedTweet.isReplied = Boolean(lt.LikedTweet.isReplied)
+        return lt
+      })
+
+      return res.json(results)
+    } catch (error) {
+      // 系統出錯
+      error.code = 500
+      return next(error)
+    }
+  },
+  // 獲取對應使用者所回覆的推文之清單(含回覆內容)
+  getReplies: async (req, res, next) => {
+    try {
+      const error = new Error()
+      const loginUserId = authHelpers.getUser(req).id
+      const targetUserId = req.params.id
+
+      // 找不到使用者可以調閱他/她的回覆推文清單
+      if (isNaN(targetUserId) || !(await User.findByPk(targetUserId))) {
+        error.code = 404
+        error.message = '對應使用者不存在'
+        return next(error)
+      }
+      // 目前使用者調閱另一個使用者X的回覆推文清單
+      // (含暱稱/名稱、帳號、頭像、封面、它所回覆的推文是否也被目前使用者喜歡和回覆)
+      // (時間、推文內容等)
+      // 按回覆紀錄排序，由新至舊
+      const findOption = {
+        where: {
+          UserId: targetUserId
+        },
+        include: [
+          {
+            model: User,
+            as: 'ReplyAuthor',
+            attributes: [
+              'id', 'account', 'name',
+              'cover', 'avatar'
+            ]
+          },
+          {
+            model: Tweet,
+            as: 'TargetTweet',
+            include: [
+              {
+                model: User,
+                as: 'TweetAuthor',
+                attributes: ['id', 'name', 'account', 'avatar', 'cover']
+              }
+            ]
+          }
+        ],
+        attributes: [
+          'id', 'UserId', 'TweetId',
+          'comment', 'createdAt', 'updatedAt',
+          [
+            sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Likes
+                      WHERE UserId = ${loginUserId} 
+                      AND TweetId = TargetTweet.id
+                    )
+            `),
+            'isLikedTweet'
+          ],
+          [
+            sequelize.literal(`
+                  EXISTS (
+                      SELECT 1 FROM Replies
+                      WHERE UserId = ${loginUserId} 
+                      AND TweetId = TargetTweet.id
+                    )
+            `),
+            'isRepliedTweet'
+          ]
+        ],
+        order: [['createdAt', 'DESC']]
+      }
+      const replies = await Reply.findAll(findOption)
+
+      const results = replies.map(r => {
+        r = r.toJSON()
+        r.isLikedTweet = Boolean(r.isLikedTweet)
+        r.isRepliedTweet = Boolean(r.isRepliedTweet)
+        return r
+      })
+
+      return res
+        .status(200)
+        .json(results)
+    } catch (error) {
+      // 系統出錯
+      error.code = 500
+      return next(error)
+    }
+  }
 }
 
 module.exports = userController
